+++ public/api-docs/v1/swagger.yaml
@@ -343,3 +343,133 @@ paths:
             '["application/json"]':
               schema:
                 "$ref": "#/components/schemas/ErrorModelV1"
+  "/v1/publications/{id}/grants":
+    get:
+      summary: Retrieve a publication's grants
+      description: Returns grant data associated with a publication
+      operationId: findPublicationGrants
+      tags:
+      - - publication
+      parameters:
+      - name: id
+        in: path
+        description: ID of publication to retrieve grants
+        required: true
+        schema:
+          type: string
+      security:
+      - api_key: []
+      responses:
+        '200':
+          description: publication grants response
+          content:
+            '["application/json"]':
+              schema:
+                type: object
+                properties:
+                  data:
+                    type: array
+                    items:
+                      type: object
+                      required:
+                      - id
+                      - type
+                      - attributes
+                      properties:
+                        id:
+                          type: string
+                          example: '123'
+                          description: The ID of the object
+                        type:
+                          type: string
+                          example: grant
+                          description: The type of the object
+                        attributes:
+                          type: object
+                          required:
+                          - title
+                          - agency
+                          - abstract
+                          - amount_in_dollars
+                          - start_date
+                          - end_date
+                          - identifier
+                          properties:
+                            title:
+                              type:
+                              - string
+                              - 'null'
+                              example: A Research Project Proposal
+                              description: The title of the grant
+                            agency:
+                              type:
+                              - string
+                              - 'null'
+                              example: National Science Foundation
+                              description: The name of the organization that awarded
+                                the grant
+                            abstract:
+                              type:
+                              - string
+                              - 'null'
+                              example: Information about this grant
+                              description: A description of the grant's purpose
+                            amount_in_dollars:
+                              type:
+                              - integer
+                              - 'null'
+                              example: 50000
+                              description: The monetary amount of the grant in U.S.
+                                dollars
+                            start_date:
+                              type:
+                              - string
+                              - 'null'
+                              example: '2017-12-05'
+                              description: The date on which the grant begins
+                            end_date:
+                              type:
+                              - string
+                              - 'null'
+                              example: '2019-12-05'
+                              description: The date on which the grant ends
+                            identifier:
+                              type:
+                              - string
+                              - 'null'
+                              example: '1789352'
+                              description: A code identifying the grant that is unique
+                                to the awarding agency
+                required:
+                - data
+        '401':
+          description: unauthorized
+          content:
+            '["application/json"]':
+              schema:
+                "$ref": "#/components/schemas/ErrorModelV1"
+        '404':
+          description: not found
+          content:
+            '["application/json"]':
+              schema:
+                "$ref": "#/components/schemas/ErrorModelV1"
+  "/v1/publications":
+    patch:
+      summary: update_all publication
+      responses:
+        '200':
+          description: successful
+  "/v1/publications/{id}":
+    parameters:
+    - name: id
+      in: path
+      description: id
+      required: true
+      schema:
+        type: string
+    get:
+      summary: show publication
+      responses:
+        '200':
+          description: successful
